+++ docs/density.md
@@ -1,0 +1,1 @@
+<!-- ::: rbig.density.bin_estimation -->
