--- rbig/__init__.py
@@ -1,1 +1,0 @@
-from .rbig import RBIG, RBIGMI, RBIGKLD
