--- rbig/model/__init__.py
@@ -1,6 +1,0 @@
-from rbig.model._rbig import RBIG
-
-
-__all__ = [
-    "RBIG",
-]
